configure: Do not treat warnings as errors by default
When the project was built from a cloned Git repository, the option
"--enable-werror" was enabled, which caused the build to abort on
warnings. Because this is an relatively old QEMU version, a number of
warnings occur with newer compilers. To be able to build the project
despite of this, now "--disable-werror" is the standard.